% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-fit_models.R
\name{create_markovchains}
\alias{create_markovchains}
\title{Fits markov chains and builds conditions}
\usage{
create_markovchains(dt)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{set_conditions}}}
}
\value{
The input \code{data.table} with an extra variable 'mcs' a markov chain sequence of
weather conditions
}
\description{
Fit and create markov chains for daily weather station conditions
}
\details{
Probabilistic markov chain sequences are learnt from weather conditions
for each month at each station found in the input data.
}

