% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-format_output.R
\name{format_output}
\alias{format_output}
\title{Tidy output for pretty printing}
\usage{
format_output(dt)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{generate_metrics}}}
}
\value{
A \code{data.table} formatted according to the specifications in details
}
\description{
Format seaweed data according to specifications
}
\details{
Synthetic Australian weather data is fromatted as such:

station location time condition temperature pressure humidity\cr
SYD|-33.86,151.21,39|2015-12-23T05:02:12Z|Rain|+12.5|1004.3|97\cr
MEL|-37.83,144.98,7|2015-12-24T15:30:55Z|Snow|-5.3|998.4|55\cr
ADL|-34.92,138.62,48|2016-01-03T12:35:37Z|Sunny|+39.4|1114.1|12\cr

A random time component is added to the date variable to align to the specs.
}
\seealso{
\code{\link{create_markovchains}} \code{\link{compute_pdfs}}
}

