% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-compute_numerics.R
\name{generate_metrics}
\alias{generate_metrics}
\title{Pick daily temperature, pressure and humidity from a normal distribution}
\usage{
generate_metrics(dt, pdfs)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{create_markovchains}}}

\item{pdfs}{A \code{data.table} created by the function \code{\link{compute_pdfs}}}
}
\value{
A \code{data.table} with temperature, pressure and humidity calculated
from a random variable drawn from a normal distribution described by a
}
\description{
Pick daily temperature, pressure and humidity from a normal distribution
}
\details{
Temperature, pressure and humidity are randomly chosen from a normal
distribution described by the mean and standard deviation (the probability density
function) for daily weather conditions observed at each station over each month.
}
\examples{
\dontrun{head(generate_metrics(dt, pdfs))}
station       Date year month day condition temperature pressure humidity
HBA 2016-09-01 2016     9   1    Cloudy        11.1   1011.8       60
HBA 2016-09-02 2016     9   2      Rain         9.1   1024.4       75
HBA 2016-09-03 2016     9   3      Rain        10.0   1002.6       78
HBA 2016-09-04 2016     9   4      Rain         9.5    994.9       79
HBA 2016-09-05 2016     9   5     Sunny        10.1   1012.5       51
HBA 2016-09-06 2016     9   6      Rain        13.7   1002.4       67
}
\seealso{
\code{\link{create_markovchains}} \code{\link{compute_pdfs}}
}

