% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-define_conditions.R
\name{set_conditions}
\alias{set_conditions}
\title{Define weather conditions}
\usage{
set_conditions(dt)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{get_weather}}}
}
\value{
A \code{data.table} of all the input data including 4 extra variables -
\code{year}, \code{month}, \code{day} and \code{condition}
}
\description{
Defines daily weather conditions from rules applied to historical weather data.
}
\details{
Rules to determine weather condition are based on commonsense and are
based upon mean values of "CloudCover" and "Precipitation" > 0. Weather onditions
can take only one of three values - "Sunny", "Cloudy" or "Rainy".

The function splits the \code{Date} "YYYY-MM-DD" into "year", "month" and "day"
for use in the functions \code{create_markovchains(dt)} and \code{compute_pdfs(dt)}
that are used to model the weather data.
}
\seealso{
\code{\link{create_markovchains}} \code{\link{compute_pdfs}}
}

