% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-write_output.R
\name{write_output}
\alias{write_output}
\title{Write generated weather data to file}
\usage{
write_output(dt)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{format_output}}}
}
\description{
Write a pipe delimeted file to the working directory
}
\details{
The synthetic Australian weather data generated upstream is writted out
to a pipe delimited file in the project root directory. The file is named
"seaweed_generated.dat" and it's extension is .dat. The file's data layout conforms
to the format applied by the function \code{\link{format_output}}.
}

