% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-fit_models.R
\name{compute_pdfs}
\alias{compute_pdfs}
\title{Calculate probability density functions of numeric weather characteristics}
\usage{
compute_pdfs(dt)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{create_markovchains}}}
}
\value{
A \code{data.table} of calculated means and standard deviations
}
\description{
Calculate probability density functions of numeric weather characteristics
}
\details{
Means and standard deviations are calculated for each weather condition's
observed temperature, pressure and humidity over the month for each station.
In the event of no monthly variation, a mertic is imputed from the complete
range of observations.
}
\examples{
\dontrun{head(compute_pdfs(dt))}
condition station month t_mean  t_sd   p_mean     p_sd   h_mean     h_sd
Cloudy   HBA     9 11.35714 2.239751 1014.429 5.079586 69.35714 9.919733
Rain     HBA     9 11.10000 2.330951 1006.000 7.630349 73.60000 9.788883
Sunny    HBA     9 11.66667 1.505545 1012.667 9.521905 60.50000 9.731393
Cloudy   MEL     9 12.11111 1.536591 1017.333 8.000000 69.88889 4.512329
Rain     MEL     9 11.10000 1.071153 1010.800 8.401754 78.20000 8.294577
Sunny    MEL     9 17.00000 5.000000 1019.000 7.000000 61.00000 9.000000
}

