% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-format_output.R
\name{format_output}
\alias{format_output}
\title{Tidy output for pretty printing}
\usage{
format_output(dt)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{generate_metrics}}}
}
\value{
A \code{data.table} formatted according to the specifications in details
}
\description{
Format seaweed data according to specifications
}
\details{
Synthetic Australian weather data is fromatted as such:

station location time condition temperature pressure humidity\cr
SYD|-33.86,151.21,39|2015-12-23T05:02:12Z|Rain|+12.5|1004.3|97\cr
MEL|-37.83,144.98,7|2015-12-24T15:30:55Z|Snow|-5.3|998.4|55\cr
ADL|-34.92,138.62,48|2016-01-03T12:35:37Z|Sunny|+39.4|1114.1|12\cr

A random time component is added to the date variable to align to the specs.
}
\examples{
\dontrun{head(format_output(dt))}
station          location                 time condition temperature pressure humidity
ADL   -34.57,138.31,4 2016-09-01T10:09:34Z      Rain        12.1   1021.6       70
ASP -23.48,133.54,541 2016-09-01T10:16:28Z    Cloudy        21.7   1014.8       43
BNE   -27.23,153.08,5 2016-09-01T18:50:40Z     Sunny        19.1   1013.0       54
CBR -35.17,149.10,577 2016-09-01T13:40:24Z      Rain        10.1   1012.2       78
CNS   -16.52,145.45,7 2016-09-01T21:33:33Z      Rain        22.7   1010.4       74
DRW  -12.25,130.54,30 2016-09-01T08:46:37Z     Sunny        28.9   1010.1       51
}
\seealso{
\code{\link{create_markovchains}} \code{\link{compute_pdfs}}
}

