% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-compute_numerics.R
\name{generate_metrics}
\alias{generate_metrics}
\title{Random sampling from statistical models}
\usage{
generate_metrics(dt, pdfs)
}
\arguments{
\item{dt}{A \code{data.table} created by the function \code{\link{create_markovchains}}}

\item{pdfs}{A \code{data.table} created by the function \code{\link{compute_pdfs}}}
}
\value{
A \code{data.table} with temperature, pressure and humidity calculated
from a random variable drawn from a normal distribution described by a
}
\description{
Pick daily temperature, pressure and humidity from a normal distribution
}
\details{
Temperature, pressure and humidity are randomly chosen from a normal
distribution described by the mean and standard deviation (the probability density
function) for daily weather conditions observed at each station over each month.
}
\seealso{
\code{\link{create_markovchains}} \code{\link{compute_pdfs}}
}

